+++ CHANGELOG.md
@@ -29,3 +29,4 @@
 
 - Show comments in PDF grid sections
 - Add markdown support for verse text
+- Add font templates to the settings